tidy up some stuff after review
